+++ apps/doc/tests/e2e/home.spec.ts
@@ -5,5 +5,7 @@ test('index page has expected headings', async ({page}) => {
 	await expect(
 		page.getByRole('heading', {name: 'Fat Fuzzy Rocks'}),
 	).toBeVisible()
-	await expect(page.getByRole('heading', {name: '✨ Highlights'})).toBeVisible()
+	await expect(
+		page.getByRole('heading', {name: '✨ Contents ✨'}),
+	).toBeVisible()
 })