+++ packages/playbook/src/lib/fixtures/js/layouts.ts
@@ -4,12 +4,5 @@ export const content = {
 }
 
 export default {
-	Reveal: {
-		slug: 'ui',
-		path: '/ui',
-		title: 'Reveal',
-		formaction: 'updateState',
-		actionPath: '/',
-	},
 	content,
 }