+++ packages/style/src/lib/css/recipes/index.js
@@ -4,9 +4,11 @@ import './nav.css' /* Use for navigation (navigation = links = hrefs = <a>) */
 import './menu.css' /* Use for menus (menus = actions -> use forms [for no-js] -> buttons = <button>) and derived elements */
 import './tabs.css' /* Use for tabs */
 import './context.css' /* Use for header context box */
+import './scrolly.css' /* Use with ScrollReveal component  */
 
 import './settings.css' /* A Settings menu component (for user input to adjust brightness, contrast, language, ...) */
 import './header-app.css' /* Use for main app header */
+import './header-settings.css' /* Use for home page settings header */
 import './pages.css' /* Define page layout varians here */
 import './footer.css' /* Use for main app footer */
 import './article.css' /* Use for articles (header + section(s) + optional footer) */